Extract out field repository module
