+++ lib/pipekit.rb
@@ -5,6 +5,7 @@ require "pipekit/request"
 require "pipekit/result"
 require "pipekit/response"
 require "pipekit/repository"
+require "pipekit/field_repository"
 require "pipekit/person"
 require "pipekit/deal"
 require "pipekit/person_field"
